backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
asyncpg==0.30.0
redis==5.2.1
pydantic-settings==2.7.0
httpx==0.28.1
